joy_control/requirements.txt
hidapi
awsdeepracer_control